Aborted mission to create a delete button
I created an array to store the value of each entry and have the last
value in the array popped when the delete button is clicked. The last
item in the array is also removed from the display of the calculator.

But the problem is that once the delete button is used, the calculator
throws an error. If the delete button isn't used. All works well. I will
put a pin in it for now and visit it at a later time.